@@ -4,10 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
-import org.springframework.util.CollectionUtils;
 
-import javax.annotation.Resource;
-import java.util.Collection;
+import jakarta.annotation.Resource;
+import java.util.Arrays;
 import java.util.concurrent.TimeUnit;
 
 @Component("redisUtils")
@@ -23,16 +22,18 @@ public class RedisUtils<V> {
      *
      * @param key 可以传一个值 或多个
      */
+    @SuppressWarnings("null")
     public void delete(String... key) {
         if (key != null && key.length > 0) {
             if (key.length == 1) {
                 redisTemplate.delete(key[0]);
             } else {
-                redisTemplate.delete((Collection<String>) CollectionUtils.arrayToList(key));
+                redisTemplate.delete(Arrays.asList(key));
             }
         }
     }
 
+    @SuppressWarnings("null")
     public V get(String key) {
         return key == null ? null : redisTemplate.opsForValue().get(key);
     }
@@ -44,6 +45,7 @@ public class RedisUtils<V> {
      * @param value 值
      * @return true成功 false失败
      */
+    @SuppressWarnings("null")
     public boolean set(String key, V value) {
         try {
             redisTemplate.opsForValue().set(key, value);
@@ -62,6 +64,7 @@ public class RedisUtils<V> {
      * @param time  时间(秒) time要大于0 如果time小于等于0 将设置无限期
      * @return true成功 false 失败
      */
+    @SuppressWarnings("null")
     public boolean setex(String key, V value, long time) {
         try {
             if (time > 0) {
