@@ -7,20 +7,32 @@ import com.easypan.entity.enums.ResponseCodeEnum;
 import com.easypan.entity.vo.PaginationResultVO;
 import com.easypan.entity.vo.ResponseVO;
 import com.easypan.utils.CopyTools;
-import com.easypan.utils.StringTools;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
+import jakarta.annotation.Resource;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
+import java.nio.channels.Channels;
+import java.nio.channels.FileChannel;
+import java.nio.channels.WritableByteChannel;
+import java.nio.ByteBuffer;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
 
 public class ABaseController {
 
     private static final Logger logger = LoggerFactory.getLogger(ABaseController.class);
+    private static final int STREAM_BUFFER_SIZE = 64 * 1024;
+    private static final long TRANSFER_CHUNK_SIZE = 8L * 1024 * 1024;
+
+    @Resource
+    private com.easypan.component.S3Component s3Component;
 
     protected static final String STATUC_SUCCESS = "success";
 
@@ -56,41 +68,67 @@ public class ABaseController {
     }
 
     protected void readFile(HttpServletResponse response, String filePath) {
-        if (!StringTools.pathIsOk(filePath)) {
-            return;
-        }
         OutputStream out = null;
-        FileInputStream in = null;
         try {
-            File file = new File(filePath);
-            if (!file.exists()) {
-                return;
-            }
-            in = new FileInputStream(file);
-            byte[] byteData = new byte[1024];
+            response.setBufferSize(STREAM_BUFFER_SIZE);
             out = response.getOutputStream();
-            int len = 0;
-            while ((len = in.read(byteData)) != -1) {
-                out.write(byteData, 0, len);
+            if (isLocalPath(filePath)) {
+                transferLocalFile(filePath, out);
+            } else {
+                transferStreamFromS3(filePath, out);
             }
             out.flush();
         } catch (Exception e) {
-            logger.error("读取文件异常", e);
-        } finally {
-            if (out != null) {
-                try {
-                    out.close();
-                } catch (IOException e) {
-                    logger.error("IO异常", e);
+            logger.error("Read file failed, path: {}", filePath, e);
+        }
+    }
+
+    private boolean isLocalPath(String filePath) {
+        return filePath.startsWith("/") || (filePath.contains(":") && filePath.indexOf(":") < 3);
+    }
+
+    private void transferLocalFile(String filePath, OutputStream out) throws IOException {
+        Path path = new File(filePath).toPath();
+        if (!Files.exists(path) || !Files.isRegularFile(path)) {
+            return;
+        }
+        try (FileChannel fileChannel = FileChannel.open(path, StandardOpenOption.READ)) {
+            WritableByteChannel outChannel = Channels.newChannel(out);
+            long size = fileChannel.size();
+            long position = 0L;
+            while (position < size) {
+                long transferred = fileChannel.transferTo(position, Math.min(TRANSFER_CHUNK_SIZE, size - position), outChannel);
+                if (transferred <= 0) {
+                    transferRemainingByBuffer(fileChannel, position, out);
+                    break;
                 }
+                position += transferred;
             }
-            if (in != null) {
-                try {
-                    in.close();
-                } catch (IOException e) {
-                    logger.error("IO异常", e);
-                }
+        }
+    }
+
+    private void transferRemainingByBuffer(FileChannel fileChannel, long position, OutputStream out) throws IOException {
+        fileChannel.position(position);
+        byte[] buffer = new byte[STREAM_BUFFER_SIZE];
+        ByteBuffer byteBuffer = ByteBuffer.wrap(buffer);
+        int read;
+        while ((read = fileChannel.read(byteBuffer)) != -1) {
+            out.write(buffer, 0, read);
+            byteBuffer.clear();
+        }
+    }
+
+    private void transferStreamFromS3(String filePath, OutputStream out) throws IOException {
+        try (InputStream in = s3Component.getInputStream(filePath)) {
+            if (in == null) {
+                return;
+            }
+            byte[] buffer = new byte[STREAM_BUFFER_SIZE];
+            int len;
+            while ((len = in.read(buffer)) != -1) {
+                out.write(buffer, 0, len);
             }
         }
     }
 }
+
