@@ -17,7 +17,7 @@ import com.easypan.utils.StringTools;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.Date;
 import java.util.List;
 
@@ -28,7 +28,7 @@ import java.util.List;
 public class FileShareServiceImpl implements FileShareService {
 
     @Resource
-    private FileShareMapper<FileShare, FileShareQuery> fileShareMapper;
+    private FileShareMapper fileShareMapper;
 
     /**
      * 根据条件查询列表
