@@ -8,8 +8,13 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 @Configuration
 public class SecurityConfig {
 
+    /**
+     * 密码编码器配置
+     * 使用 BCrypt 强度 12（推荐生产环境配置）
+     * 强度范围：4-31，默认 10，每增加 1 计算时间翻倍
+     */
     @Bean
     public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
+        return new BCryptPasswordEncoder(12);
     }
 }
