@@ -1,40 +1,11 @@
 package com.easypan.entity.dto;
 
+import lombok.Data;
+
+@Data
 public class DownloadFileDto {
     private String downloadCode;
     private String fileId;
     private String fileName;
     private String filePath;
-
-    public String getDownloadCode() {
-        return downloadCode;
-    }
-
-    public void setDownloadCode(String downloadCode) {
-        this.downloadCode = downloadCode;
-    }
-
-    public String getFileId() {
-        return fileId;
-    }
-
-    public void setFileId(String fileId) {
-        this.fileId = fileId;
-    }
-
-    public String getFileName() {
-        return fileName;
-    }
-
-    public void setFileName(String fileName) {
-        this.fileName = fileName;
-    }
-
-    public String getFilePath() {
-        return filePath;
-    }
-
-    public void setFilePath(String filePath) {
-        this.filePath = filePath;
-    }
 }
