@@ -1,5 +1,8 @@
 package com.easypan.entity.dto;
 
+import lombok.Data;
+
+@Data
 public class QQInfoDto {
     private Integer ret;
     private String msg;
@@ -7,52 +10,4 @@ public class QQInfoDto {
     private String figureurl_qq_1;
     private String figureurl_qq_2;
     private String gender;
-
-    public Integer getRet() {
-        return ret;
-    }
-
-    public void setRet(Integer ret) {
-        this.ret = ret;
-    }
-
-    public String getMsg() {
-        return msg;
-    }
-
-    public void setMsg(String msg) {
-        this.msg = msg;
-    }
-
-    public String getNickname() {
-        return nickname;
-    }
-
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
-
-    public String getFigureurl_qq_1() {
-        return figureurl_qq_1;
-    }
-
-    public void setFigureurl_qq_1(String figureurl_qq_1) {
-        this.figureurl_qq_1 = figureurl_qq_1;
-    }
-
-    public String getFigureurl_qq_2() {
-        return figureurl_qq_2;
-    }
-
-    public void setFigureurl_qq_2(String figureurl_qq_2) {
-        this.figureurl_qq_2 = figureurl_qq_2;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
 }
