@@ -1,27 +1,12 @@
 package com.easypan.entity.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
+import lombok.Data;
 import java.io.Serializable;
 
+@Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class UploadResultDto implements Serializable {
     private String fileId;
     private String status;
-
-    public String getFileId() {
-        return fileId;
-    }
-
-    public void setFileId(String fileId) {
-        this.fileId = fileId;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
 }
