@@ -1,19 +1,19 @@
 package com.easypan.entity.po;
 
-import com.easypan.entity.enums.DateTimePatternEnum;
-import com.easypan.utils.DateUtil;
 import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Data;
+import com.mybatisflex.annotation.Table;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serializable;
 import java.util.Date;
 
-
 /**
  * 分享信息
  */
+@Data
+@Table("file_share")
 public class FileShare implements Serializable {
-
 
     /**
      * 分享ID
@@ -59,7 +59,6 @@ public class FileShare implements Serializable {
      */
     private Integer showCount;
 
-
     private String fileName;
 
     /**
@@ -68,12 +67,12 @@ public class FileShare implements Serializable {
     private Integer folderType;
 
     /**
-     * 1:视频 2:音频  3:图片 4:文档 5:其他
+     * 1:视频 2:音频 3:图片 4:文档 5:其他
      */
     private Integer fileCategory;
 
     /**
-     * 1:视频 2:音频  3:图片 4:pdf 5:doc 6:excel 7:txt 8:code 9:zip 10:其他
+     * 1:视频 2:音频 3:图片 4:pdf 5:doc 6:excel 7:txt 8:code 9:zip 10:其他
      */
     private Integer fileType;
 
@@ -81,116 +80,4 @@ public class FileShare implements Serializable {
      * 封面
      */
     private String fileCover;
-
-    public String getFileName() {
-        return fileName;
-    }
-
-    public void setFileName(String fileName) {
-        this.fileName = fileName;
-    }
-
-    public void setShareId(String shareId) {
-        this.shareId = shareId;
-    }
-
-    public String getShareId() {
-        return this.shareId;
-    }
-
-    public void setFileId(String fileId) {
-        this.fileId = fileId;
-    }
-
-    public String getFileId() {
-        return this.fileId;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public String getUserId() {
-        return this.userId;
-    }
-
-    public void setValidType(Integer validType) {
-        this.validType = validType;
-    }
-
-    public Integer getValidType() {
-        return this.validType;
-    }
-
-    public void setExpireTime(Date expireTime) {
-        this.expireTime = expireTime;
-    }
-
-    public Date getExpireTime() {
-        return this.expireTime;
-    }
-
-    public void setShareTime(Date shareTime) {
-        this.shareTime = shareTime;
-    }
-
-    public Date getShareTime() {
-        return this.shareTime;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getCode() {
-        return this.code;
-    }
-
-    public void setShowCount(Integer showCount) {
-        this.showCount = showCount;
-    }
-
-    public Integer getShowCount() {
-        return this.showCount;
-    }
-
-    public Integer getFolderType() {
-        return folderType;
-    }
-
-    public void setFolderType(Integer folderType) {
-        this.folderType = folderType;
-    }
-
-    public Integer getFileCategory() {
-        return fileCategory;
-    }
-
-    public void setFileCategory(Integer fileCategory) {
-        this.fileCategory = fileCategory;
-    }
-
-    public Integer getFileType() {
-        return fileType;
-    }
-
-    public void setFileType(Integer fileType) {
-        this.fileType = fileType;
-    }
-
-    public String getFileCover() {
-        return fileCover;
-    }
-
-    public void setFileCover(String fileCover) {
-        this.fileCover = fileCover;
-    }
-
-    @Override
-    public String toString() {
-        return "分享ID:" + (shareId == null ? "空" : shareId) + "，文件ID:" + (fileId == null ? "空" : fileId) + "，用户ID:" + (userId == null ? "空" : userId) + "，有效期类型 0:1天 " +
-                "1:7天 2:30天 3:永久有效:" + (validType == null ? "空" : validType) + "，失效时间:" + (expireTime == null ? "空" : DateUtil.format(expireTime,
-                DateTimePatternEnum.YYYY_MM_DD_HH_MM_SS.getPattern())) + "，分享时间:" + (shareTime == null ? "空" : DateUtil.format(shareTime,
-                DateTimePatternEnum.YYYY_MM_DD_HH_MM_SS.getPattern())) + "，提取码:" + (code == null ? "空" : code) + "，浏览次数:" + (showCount == null ? "空" : showCount);
-    }
 }
