@@ -1,26 +1,27 @@
 package com.easypan.entity.po;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-import java.util.Date;
-
-import com.easypan.entity.enums.DateTimePatternEnum;
-import com.easypan.utils.DateUtil;
 import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serializable;
+import java.util.Date;
 
+import com.mybatisflex.annotation.Id;
+import com.mybatisflex.annotation.KeyType;
+import com.mybatisflex.annotation.Table;
 
 /**
  * 用户信息
  */
+@Data
+@Table(value = "user_info")
 public class UserInfo implements Serializable {
-
 
     /**
      * 用户ID
      */
+    @Id(keyType = KeyType.None)
     private String userId;
 
     /**
@@ -76,98 +77,4 @@ public class UserInfo implements Serializable {
      * 总空间单位byte
      */
     private Long totalSpace;
-
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public String getUserId() {
-        return this.userId;
-    }
-
-    public void setNickName(String nickName) {
-        this.nickName = nickName;
-    }
-
-    public String getNickName() {
-        return this.nickName;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getEmail() {
-        return this.email;
-    }
-
-    public void setQqAvatar(String qqAvatar) {
-        this.qqAvatar = qqAvatar;
-    }
-
-    public String getQqAvatar() {
-        return this.qqAvatar;
-    }
-
-    public void setQqOpenId(String qqOpenId) {
-        this.qqOpenId = qqOpenId;
-    }
-
-    public String getQqOpenId() {
-        return this.qqOpenId;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getPassword() {
-        return this.password;
-    }
-
-    public void setJoinTime(Date joinTime) {
-        this.joinTime = joinTime;
-    }
-
-    public Date getJoinTime() {
-        return this.joinTime;
-    }
-
-    public void setLastLoginTime(Date lastLoginTime) {
-        this.lastLoginTime = lastLoginTime;
-    }
-
-    public Date getLastLoginTime() {
-        return this.lastLoginTime;
-    }
-
-    public void setStatus(Integer status) {
-        this.status = status;
-    }
-
-    public Integer getStatus() {
-        return this.status;
-    }
-
-    public void setUseSpace(Long useSpace) {
-        this.useSpace = useSpace;
-    }
-
-    public Long getUseSpace() {
-        return this.useSpace;
-    }
-
-    public void setTotalSpace(Long totalSpace) {
-        this.totalSpace = totalSpace;
-    }
-
-    public Long getTotalSpace() {
-        return this.totalSpace;
-    }
-
-    @Override
-    public String toString() {
-        return "用户ID:" + (userId == null ? "空" : userId) + "，昵称:" + (nickName == null ? "空" : nickName) + "，邮箱:" + (email == null ? "空" : email) + "，qqAvatar:" + (qqAvatar == null ? "空" : qqAvatar) + "，qqOpenId:" + (qqOpenId == null ? "空" : qqOpenId) + "，密码:" + (password == null ? "空" : password) + "，加入时间:" + (joinTime == null ? "空" : DateUtil.format(joinTime, DateTimePatternEnum.YYYY_MM_DD_HH_MM_SS.getPattern())) + "，最后登录时间:" + (lastLoginTime == null ? "空" : DateUtil.format(lastLoginTime, DateTimePatternEnum.YYYY_MM_DD_HH_MM_SS.getPattern())) + "，0:禁用 1:正常:" + (status == null ? "空" : status) + "，useSpace:" + (useSpace == null ? "空" : useSpace) + "，totalSpace:" + (totalSpace == null ? "空" : totalSpace);
-    }
 }
