@@ -1,22 +1,9 @@
 package com.easypan.entity.vo;
 
+import lombok.Data;
+
+@Data
 public class FolderVO {
     private String fileName;
     private String fileId;
-
-    public String getFileName() {
-        return fileName;
-    }
-
-    public void setFileName(String fileName) {
-        this.fileName = fileName;
-    }
-
-    public String getFileId() {
-        return fileId;
-    }
-
-    public void setFileId(String fileId) {
-        this.fileId = fileId;
-    }
 }
