@@ -1,8 +1,15 @@
 package com.easypan.entity.vo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import java.util.ArrayList;
 import java.util.List;
 
-
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class PaginationResultVO<T> {
 	private Integer totalCount;
 	private Integer pageSize;
@@ -17,62 +24,17 @@ public class PaginationResultVO<T> {
 		this.list = list;
 	}
 
-    public PaginationResultVO(Integer totalCount, Integer pageSize, Integer pageNo, Integer pageTotal, List<T> list) {
-        if (pageNo == 0) {
-            pageNo = 1;
-        }
-        this.totalCount = totalCount;
-        this.pageSize = pageSize;
-        this.pageNo = pageNo;
-        this.pageTotal = pageTotal;
-        this.list = list;
-    }
-
 	public PaginationResultVO(List<T> list) {
 		this.list = list;
 	}
 
-	public PaginationResultVO() {
-
+	// Custom setter for pageNo logic or custom constructor if needed
+	// The original had a check in the 5-args constructor
+	public static <T> PaginationResultVO<T> build(Integer totalCount, Integer pageSize, Integer pageNo,
+			Integer pageTotal, List<T> list) {
+		if (pageNo == null || pageNo == 0) {
+			pageNo = 1;
+		}
+		return new PaginationResultVO<>(totalCount, pageSize, pageNo, pageTotal, list);
 	}
-
-	public Integer getTotalCount() {
-		return totalCount;
-	}
-
-	public void setTotalCount(Integer totalCount) {
-		this.totalCount = totalCount;
-	}
-
-	public Integer getPageSize() {
-		return pageSize;
-	}
-
-	public void setPageSize(Integer pageSize) {
-		this.pageSize = pageSize;
-	}
-
-	public Integer getPageNo() {
-		return pageNo;
-	}
-
-	public void setPageNo(Integer pageNo) {
-		this.pageNo = pageNo;
-	}
-
-	public List<T> getList() {
-		return list;
-	}
-
-	public void setList(List<T> list) {
-		this.list = list;
-	}
-
-	public Integer getPageTotal() {
-        return pageTotal;
-    }
-
-    public void setPageTotal(Integer pageTotal) {
-        this.pageTotal = pageTotal;
-    }
 }
