@@ -1,7 +1,8 @@
 package com.easypan.utils;
 
 
-import com.easypan.entity.enums.DateTimePatternEnum;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -12,6 +13,7 @@ import java.util.Map;
 
 public class DateUtil {
 
+    private static final Logger logger = LoggerFactory.getLogger(DateUtil.class);
     private static final Object lockObj = new Object();
     private static Map<String, ThreadLocal<SimpleDateFormat>> sdfMap = new HashMap<String, ThreadLocal<SimpleDateFormat>>();
 
@@ -43,7 +45,7 @@ public class DateUtil {
         try {
             return getSdf(pattern).parse(dateStr);
         } catch (ParseException e) {
-            e.printStackTrace();
+            logger.error("日期解析失败: dateStr={}, pattern={}", dateStr, pattern, e);
         }
         return new Date();
     }
@@ -52,9 +54,5 @@ public class DateUtil {
         Calendar calendar = Calendar.getInstance();
         calendar.add(Calendar.DAY_OF_YEAR, day);
         return calendar.getTime();
-    }
-
-    public static void main(String[] args) {
-        System.out.println(format(getAfterDate(1), DateTimePatternEnum.YYYY_MM_DD_HH_MM_SS.getPattern()));
     }
 }
