@@ -6,7 +6,7 @@
 
 <script setup>
 import { reactive } from "vue";
-import zhCn from "element-plus/lib/locale/lang/zh-cn";
+import zhCn from "element-plus/es/locale/lang/zh-cn";
 const locale = zhCn;
 const config = reactive({
   max: 1,
